+++ src/Transactions/TransactionsList/TransactionsFilters/ExpenseClassFilter/index.js
@@ -1,0 +1,1 @@
+export { default as ExpenseClassFilter } from './ExpenseClassFilterContainer';