+++ src/common/utils/getExpenseClassesForSelect.js
@@ -1,0 +1,6 @@
+export const getExpenseClassesForSelect = (expenseClasses = []) => (
+  expenseClasses.map(({ id, name }) => ({
+    label: name,
+    value: id,
+  }))
+);