+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change history for ui-finance
 
 ## (IN PROGRESS)
+
+* Display and filter by expense class in transaction view. Refs UIF-241
 * Display "Net transfers" amount on Ledger view. Refs UIF-236
 * Allowable expenditure/encumbrance display as 0% when not set. Refs UIF-240
 * Edit fiscal year - Form elements must have labels. Refs UIF-197